requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==2.2.11
django-classy-tags==1.0.0
django-cms==3.7.1
django-entangled==0.3
django-filer==1.7.0
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-text-ckeditor==3.9.0
easy-thumbnails==2.7
html5lib==1.0.1
idna==2.9
jsonfield==3.1.0
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.8
webencodings==0.5.1
tox
djangocms-plus